feat: implement account validation and recovery on frontend
commit a9f589548ef8b100046bd44960392e3d03105033
Author: anagrahovac <grahovac.ana99@gmail.com>
Date:   Fri May 20 00:05:34 2022 +0200

    feat: add account recovery component

commit 6ea7e4acb678f04e80e2caa0f6b03a21c1304174
Author: anagrahovac <grahovac.ana99@gmail.com>
Date:   Thu May 19 23:22:11 2022 +0200

    style: rename component

commit e1fcc734226ec522a154f48c8e5a1c1ac19acb13
Author: anagrahovac <grahovac.ana99@gmail.com>
Date:   Thu May 19 21:21:16 2022 +0200

    feat: implement password reset on frontend

commit f5372702c910007e55d8a5a99a79f916cac2c287
Author: anagrahovac <grahovac.ana99@gmail.com>
Date:   Thu May 19 18:37:53 2022 +0200

    feat: add error page

commit fb7458472ee42978360d7782130184e9d9379919
Author: anagrahovac <grahovac.ana99@gmail.com>
Date:   Thu May 19 17:50:06 2022 +0200

    feat: implement account validation on frontend

+++ pki-app/src/main/java/bsep/pkiapp/service/AuthenticationService.java
@@ -93,14 +93,18 @@ public class AuthenticationService {
         return tokenUtils.getRoleFromToken(token);
     }
 
-    public void confirmAccount(String token) {
+    public boolean confirmAccount(String token) {
         ConfirmationToken confirmationToken = confirmationTokenService.findByToken(token);
-        if (confirmationToken.getTokenType().equals(ConfirmationTokenType.REGISTRATION_TOKEN)) {
+        if (confirmationToken != null && confirmationToken.getTokenType().equals(ConfirmationTokenType.REGISTRATION_TOKEN)) {
             User user = userService.getByEmail(confirmationToken.getEmail());
+            if(user == null)
+                return false;
             user.setActivated(true);
             userService.saveUser(user);
             confirmationTokenService.deleteToken(confirmationToken);
+            return true;
         }
+        return false;
     }
 
     public void recoverAccount(String email) {
@@ -123,5 +127,13 @@ public class AuthenticationService {
             userService.saveUser(user);
             confirmationTokenService.deleteToken(confirmationToken);
         }
+    }
+
+    public boolean isTokenValid(String token) {
+        ConfirmationToken confirmationToken = confirmationTokenService.findByToken(token);
+        if (confirmationToken != null) {
+            return true;
+        }
+        return false;
     }
 }
